+++ CHANGELOG.md
@@ -3,8 +3,15 @@ All notable changes to this project will be documented in this file.
 
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v0.2.0] - 2020-03-29
+### Added
+- :sparkle: Adds more info popup
+- :sparkle: Adds app version number to footer
+- :arrow_up: Updated dependencies
+
 ## [v0.1.0] - 2020-02-12
 ### Added
 - :tada: initial app
 
+[v0.2.0]: https://github.com/cujarrett/oil-countdown/compare/v0.1.0...v0.2.0
 [v0.1.0]: https://github.com/cujarrett/oil-countdown/releases/tag/v0.1.0